refactor(files): split file_management_service into focused modules
Extracts FileValidationService, FileInfoService, FileBackupService +
FileOperationService, and FileSearchService from the monolithic
app/files/application/management.py (1,309 lines) into dedicated
modules under app/files/application/. The FileManagementService
façade and the module-level singleton remain in management.py.

Backwards-compatible: all moved symbols are re-exported from
management.py so existing imports and test patches continue to work.
No behavior, signature, or error-message changes.

Resolves #156
