+++ app/files/application/management.py
@@ -1,726 +1,24 @@
 import logging
-import shutil
-import zipfile
-from datetime import datetime
 from pathlib import Path
-from typing import Annotated, Any, ClassVar, Dict, List, Optional
+from typing import Any, Dict, List, Optional
 
-import aiofiles
 from fastapi import UploadFile
 from sqlalchemy.orm import Session
 
 from app.core.config import settings
 from app.core.exceptions import (
-    AccessDeniedException,
     FileAlreadyExistsError,
-    FileMissingError,
-    FileOperationException,
-    FileTooLargeError,
-    InvalidFileTypeError,
     InvalidRequestException,
-    ServerNotFoundException,
-    handle_file_error,
 )
-from app.files.adapters.legacy import file_history_service
 from app.files.application.encoding_handler import EncodingHandler
-from app.servers.adapters.read_port import SqlAlchemyServerReadPort
+from app.files.application.file_info import FileInfoService
+from app.files.application.file_io import FileBackupService, FileOperationService
+from app.files.application.file_search import FileSearchService
+from app.files.application.path_validation import FileValidationService
 from app.types import FileType
-from app.users.domain.value_objects import Role
 from app.users.models import User
 
 logger = logging.getLogger(__name__)
-
-
-class FileValidationService:
-    """Service for validating file operations and access.
-
-    This service handles all validation logic for file operations including
-    server existence, path safety, file permissions, and access control.
-    """
-
-    def __init__(self) -> None:
-        self.allowed_extensions = {
-            "config": [".properties", ".yml", ".yaml", ".json", ".txt", ".conf"],
-            "world": [".dat", ".dat_old", ".mca", ".mcr"],
-            "plugin": [".jar"],
-            "mod": [".jar"],
-            "log": [".log", ".gz"],
-        }
-        self.restricted_files = [
-            # Core server files
-            "server.jar",
-            "eula.txt",
-            # Permission and access control files
-            "ops.json",
-            "whitelist.json",
-            "banned-players.json",
-            "banned-ips.json",
-            # Server configuration files (Bukkit/Spigot/Paper)
-            "bukkit.yml",
-            "spigot.yml",
-            "paper.yml",
-            "paper-global.yml",
-            "paper-world-defaults.yml",
-            # Plugin and command configuration
-            "plugins.yml",
-            "commands.yml",
-            "permissions.yml",
-            "help.yml",
-            # World data files (critical for world integrity)
-            "level.dat",
-            "level.dat_old",
-            "session.lock",
-            # User cache and security files
-            "usercache.json",
-            "usernamecache.json",
-            # Additional server implementation JARs
-            "minecraft_server.jar",
-            "forge.jar",
-            "fabric-server-launch.jar",
-            # Plugin management files
-            "plugin.yml",
-            "mod.toml",
-            # Proxy server configurations (for multi-server setups)
-            "config.yml",
-            "velocity.toml",
-            "waterfall.yml",
-        ]
-
-    async def validate_server_exists(
-        self,
-        server_id: Annotated[int, "ID of the server to validate"],
-        db: Annotated[Session, "Database session for queries"],
-    ) -> Annotated[Any, "Validated server view with `directory_path`"]:
-        """Validate that a server exists in the database.
-
-        Returns a view exposing `directory_path` so callers that need only
-        the on-disk path continue to work unchanged. Resolution goes
-        through `SqlAlchemyServerReadPort.get`, which returns a
-        domain-pure `ServerEntity` (soft-deleted rows are excluded).
-
-        Raises:
-            ServerNotFoundException: If server doesn't exist.
-        """
-        server = await SqlAlchemyServerReadPort(db).get(server_id)
-        if server is None:
-            raise ServerNotFoundException(str(server_id))
-        return server
-
-    def validate_server_directory(self, server_path: Path) -> None:
-        """Validate server directory exists, create if it doesn't"""
-        if not server_path.exists():
-            # Create the server directory if it doesn't exist
-            try:
-                server_path.mkdir(parents=True, exist_ok=True)
-                logger.info(f"Created server directory: {server_path}")
-            except Exception as e:
-                raise FileOperationException(
-                    "access", str(server_path), f"Failed to create server directory: {e}"
-                )
-
-    def validate_path_safety(self, server_path: Path, target_path: Path) -> None:
-        """Validate path is safe and within server directory"""
-        if not self._is_safe_path(server_path, target_path):
-            raise AccessDeniedException("file", "access")
-
-    def validate_path_exists(self, target_path: Path) -> None:
-        """Validate target path exists"""
-        if not target_path.exists():
-            raise FileMissingError("access", str(target_path))
-
-    def validate_file_readable(self, file_path: Path) -> None:
-        """Validate file is readable"""
-        if file_path.is_dir():
-            raise InvalidFileTypeError(
-                "read",
-                str(file_path),
-                "Path is a directory, not a file",
-                detected_type="directory",
-            )
-
-        if not self._is_readable_file(file_path):
-            # ``AccessDeniedException`` (403) preserved for backwards
-            # compatibility with existing tests; the policy denial is
-            # distinct from OS-level permission errors, which are mapped
-            # to :class:`FileAccessDeniedError` by ``handle_file_error``.
-            raise AccessDeniedException("file", "read")
-
-    def validate_file_writable(self, file_path: Path, user: User) -> None:
-        """Validate file can be written"""
-        if self._is_restricted_file(file_path) and user.role != Role.admin:
-            raise AccessDeniedException("file", "write")
-
-        if not self._is_writable_file(file_path):
-            raise AccessDeniedException("file", "edit")
-
-    def validate_path_deletable(self, path: Path, user: User = None) -> None:
-        """Validate path (file or directory) can be deleted"""
-        # Only validate that the path exists - no file type restrictions
-        # The actual existence check is done by validate_path_exists
-        # This method is kept for API consistency but doesn't add restrictions
-        pass
-
-    def _is_safe_path(self, server_path: Path, target_path: Path) -> bool:
-        """Check if target path is within server directory"""
-        try:
-            target_path.resolve().relative_to(server_path.resolve())
-            return True
-        except ValueError:
-            return False
-
-    def _is_readable_file(self, file_path: Path) -> bool:
-        """Check if file type is readable"""
-        suffix = file_path.suffix.lower()
-        for _, extensions in self.allowed_extensions.items():
-            if suffix in extensions:
-                return True
-        return suffix in [
-            ".txt",
-            ".md",
-            ".yml",
-            ".yaml",
-            ".json",
-            ".properties",
-            ".sh",
-            ".bat",
-            ".ini",
-            ".cfg",
-            ".xml",
-        ]
-
-    def _is_writable_file(self, file_path: Path) -> bool:
-        """Check if file type is writable"""
-        if file_path.is_dir():
-            return False
-
-        suffix = file_path.suffix.lower()
-        writable_extensions = [".properties", ".yml", ".yaml", ".json", ".txt", ".conf"]
-        return suffix in writable_extensions
-
-    def _is_restricted_file(self, file_path: Path) -> bool:
-        """Check if file is restricted from modification"""
-        return file_path.name in self.restricted_files
-
-
-class FileInfoService:
-    """Service for retrieving file information"""
-
-    def __init__(self, validation_service: FileValidationService):
-        self.validation_service = validation_service
-
-    async def get_file_info(self, file_path: Path, server_path: Path) -> Dict[str, Any]:
-        """Get comprehensive file information"""
-        try:
-            stats = file_path.stat()
-            relative_path = file_path.relative_to(server_path)
-
-            return {
-                "name": file_path.name,
-                "path": str(relative_path),
-                "size": stats.st_size if file_path.is_file() else 0,
-                "modified": datetime.fromtimestamp(stats.st_mtime).isoformat(),
-                "is_directory": file_path.is_dir(),
-                "is_file": file_path.is_file(),
-                "extension": file_path.suffix,
-                "type": self._determine_file_type(file_path),
-                "readable": self._is_file_readable(file_path),
-                "writable": self._is_file_writable(file_path),
-                "permissions": {
-                    "read": self._is_file_readable(file_path),
-                    "write": self._is_file_writable(file_path),
-                    "execute": file_path.is_file()
-                    and file_path.suffix in [".sh", ".bat", ".exe"],
-                },
-            }
-        except Exception as e:
-            handle_file_error("get info", str(file_path), e)
-
-    def _determine_file_type(self, file_path: Path) -> str:
-        """Determine file type based on extension"""
-        if file_path.is_dir():
-            return "directory"
-
-        suffix = file_path.suffix.lower()
-
-        # Text files
-        text_extensions = [
-            ".txt",
-            ".md",
-            ".yml",
-            ".yaml",
-            ".json",
-            ".properties",
-            ".conf",
-            ".log",
-            ".sh",
-            ".bat",
-            ".xml",
-            ".html",
-            ".css",
-            ".js",
-            ".py",
-            ".java",
-            ".cpp",
-            ".c",
-            ".h",
-            ".ini",
-            ".cfg",
-        ]
-
-        # Binary files
-        binary_extensions = [
-            ".jar",
-            ".zip",
-            ".tar",
-            ".gz",
-            ".exe",
-            ".dll",
-            ".so",
-            ".dylib",
-            ".png",
-            ".jpg",
-            ".jpeg",
-            ".gif",
-            ".bmp",
-            ".ico",
-            ".pdf",
-            ".dat",
-            ".mca",
-            ".mcr",
-            ".bin",
-        ]
-
-        if suffix in text_extensions:
-            return "text"
-        elif suffix in binary_extensions:
-            return "binary"
-        else:
-            return "other"
-
-    def _is_file_readable(self, file_path: Path) -> bool:
-        """Check if file is readable"""
-        if file_path.is_dir():
-            return True
-
-        suffix = file_path.suffix.lower()
-        readable_extensions = [
-            ".txt",
-            ".md",
-            ".yml",
-            ".yaml",
-            ".json",
-            ".properties",
-            ".conf",
-            ".log",
-            ".sh",
-            ".bat",
-            ".ini",
-            ".cfg",
-            ".xml",
-        ]
-        return suffix in readable_extensions
-
-    def _is_file_writable(self, file_path: Path) -> bool:
-        """Check if file is writable"""
-        return self.validation_service._is_writable_file(
-            file_path
-        ) and not self.validation_service._is_restricted_file(file_path)
-
-
-class FileBackupService:
-    """Service for creating file backups"""
-
-    async def create_file_backup(self, file_path: Path) -> str:
-        """Create backup of existing file"""
-        try:
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            backup_name = f"{file_path.name}.backup_{timestamp}"
-            backup_path = file_path.parent / backup_name
-
-            shutil.copy2(file_path, backup_path)
-            return str(backup_path)
-
-        except Exception as e:
-            handle_file_error("backup", str(file_path), e)
-
-
-class FileOperationService:
-    """Service for file system operations"""
-
-    def __init__(self, backup_service: FileBackupService):
-        self.backup_service = backup_service
-
-    async def read_file_content(
-        self, file_path: Path, encoding: str = None
-    ) -> tuple[str, str]:
-        """Read file content with automatic encoding detection
-
-        Returns:
-            Tuple of (content, detected_encoding)
-        """
-        try:
-            if encoding:
-                # If encoding is specified, use it directly
-                async with aiofiles.open(file_path, mode="r", encoding=encoding) as f:
-                    content = await f.read()
-                    return content, encoding
-            else:
-                # Use encoding detection for better compatibility
-                result = EncodingHandler.safe_read_text_file(str(file_path))
-                if result["success"]:
-                    logger.info(
-                        f"File read successfully with encoding: {result['encoding']}"
-                    )
-                    return result["content"], result["encoding"]
-                else:
-                    raise InvalidRequestException(
-                        f"Unable to decode file: {result['error']}"
-                    )
-        except UnicodeDecodeError:
-            raise InvalidRequestException(
-                f"Unable to decode file with {encoding} encoding"
-            )
-        except Exception as e:
-            handle_file_error("read", str(file_path), e)
-
-    async def read_image_as_base64(self, file_path: Path) -> str:
-        """Read image file and return as base64 encoded string"""
-        import base64
-
-        try:
-            async with aiofiles.open(file_path, mode="rb") as f:
-                image_data = await f.read()
-                return base64.b64encode(image_data).decode("utf-8")
-        except Exception as e:
-            handle_file_error("read image", str(file_path), e)
-
-    async def write_file_content(
-        self,
-        file_path: Path,
-        content: str,
-        db: Session,
-        encoding: str = "utf-8",
-        create_backup: bool = True,
-        server_id: int = None,
-        user_id: int = None,
-        description: str = None,
-    ) -> Optional[str]:
-        """Write content to file with specified encoding and create history backup
-
-        Args:
-            file_path: Path to the file to write
-            content: Content to write to the file
-            encoding: File encoding (default: utf-8)
-            create_backup: Whether to create a backup before writing
-            server_id: Server ID for backup tracking
-            user_id: User ID for backup tracking
-            description: Description for the backup
-            db: Database session (required for security-critical operations)
-
-        Returns:
-            Path to backup file if created, None otherwise
-        """
-        # Validate database session for security-critical operations
-        if db is None:
-            raise InvalidRequestException(
-                "Database session is required for file write operations"
-            )
-
-        try:
-            file_path.parent.mkdir(parents=True, exist_ok=True)
-
-            backup_record = None
-            if create_backup and file_path.exists() and server_id is not None:
-                # Read current content for history backup
-                try:
-                    async with aiofiles.open(file_path, mode="r", encoding=encoding) as f:
-                        current_content = await f.read()
-
-                    # Extract relative file path from server directory
-                    relative_path = await self._extract_relative_path(
-                        file_path, server_id, db
-                    )
-
-                    # Create history backup using new service
-                    backup_record = await file_history_service.create_version_backup(
-                        server_id=server_id,
-                        file_path=relative_path,
-                        content=current_content,
-                        user_id=user_id,
-                        description=description,
-                        db=db,
-                    )
-                except Exception as e:
-                    logger.warning(f"Failed to create history backup: {e}")
-
-            # Write new content
-            async with aiofiles.open(file_path, mode="w", encoding=encoding) as f:
-                await f.write(content)
-
-            return str(backup_record.backup_file_path) if backup_record else None
-
-        except Exception as e:
-            handle_file_error("write", str(file_path), e)
-
-    async def _extract_relative_path(
-        self, file_path: Path, server_id: int, db: Session
-    ) -> str:
-        """Extract relative path from server directory"""
-        directory_path = await SqlAlchemyServerReadPort(db).get_directory_path(server_id)
-        if directory_path is None:
-            raise ServerNotFoundException(f"Server {server_id} not found")
-
-        server_path = Path(directory_path)
-        try:
-            relative_path = file_path.relative_to(server_path)
-            return str(relative_path)
-        except ValueError:
-            # If file is not within server directory, use the filename
-            return file_path.name
-
-    # 64 KiB chunk size matches the default disk page size for most
-    # filesystems and keeps peak per-request memory bounded while
-    # streaming the upload to disk.
-    _UPLOAD_CHUNK_BYTES: ClassVar[int] = 64 * 1024
-
-    async def upload_file(self, file: UploadFile, target_path: Path) -> int:
-        """Upload file to target path and return file size.
-
-        Reads the upload in fixed-size chunks (#341) so the full payload
-        never materialises in memory. The running total is checked
-        against ``settings.FILE_MAX_UPLOAD_BYTES`` after every chunk and
-        a :class:`FileTooLargeError` is raised the moment the limit is
-        exceeded — any partial output is removed before the exception
-        propagates. ``FILE_MAX_UPLOAD_BYTES = 0`` disables enforcement.
-        """
-        from app.core.concurrency import get_semaphores
-
-        async with get_semaphores().file_io:
-            return await self._upload_file_inner(file, target_path)
-
-    async def _upload_file_inner(self, file: UploadFile, target_path: Path) -> int:
-        max_bytes = settings.FILE_MAX_UPLOAD_BYTES
-        enforce_limit = max_bytes > 0
-        try:
-            target_path.parent.mkdir(parents=True, exist_ok=True)
-
-            total = 0
-            try:
-                async with aiofiles.open(target_path, mode="wb") as f:
-                    while True:
-                        chunk = await file.read(self._UPLOAD_CHUNK_BYTES)
-                        if not chunk:
-                            break
-                        total += len(chunk)
-                        if enforce_limit and total > max_bytes:
-                            raise FileTooLargeError(
-                                "upload",
-                                str(target_path),
-                                size_bytes=total,
-                                max_bytes=max_bytes,
-                            )
-                        await f.write(chunk)
-            except FileTooLargeError:
-                try:
-                    if target_path.exists():
-                        target_path.unlink()
-                except OSError:
-                    logger.debug(
-                        "Failed to cleanup partial upload at %s",
-                        target_path,
-                        exc_info=True,
-                    )
-                raise
-
-            return total
-
-        except Exception as e:
-            handle_file_error("upload", str(target_path), e)
-
-    def delete_file_or_directory(self, path: Path) -> str:
-        """Delete file or directory and return operation type"""
-        try:
-            if path.is_file():
-                path.unlink()
-                return "file"
-            elif path.is_dir():
-                shutil.rmtree(path)
-                return "directory"
-            else:
-                # Builtin ``FileNotFoundError`` is intercepted by
-                # ``handle_file_error`` and re-raised as the structured
-                # :class:`FileMissingError` (404).
-                raise FileNotFoundError(f"Path does not exist: {path}")
-        except Exception as e:
-            handle_file_error("delete", str(path), e)
-
-    def create_directory(self, path: Path) -> None:
-        """Create directory"""
-        try:
-            path.mkdir(parents=True, exist_ok=True)
-        except Exception as e:
-            handle_file_error("create directory", str(path), e)
-
-    def move_file_or_directory(self, source: Path, destination: Path) -> None:
-        """Move file or directory"""
-        try:
-            destination.parent.mkdir(parents=True, exist_ok=True)
-            shutil.move(str(source), str(destination))
-        except Exception as e:
-            handle_file_error("move", f"{source} to {destination}", e)
-
-    def extract_archive(self, archive_path: Path, extract_to: Path) -> List[str]:
-        """Extract archive and return list of extracted files"""
-        try:
-            extracted_files = []
-
-            if archive_path.suffix.lower() == ".zip":
-                with zipfile.ZipFile(archive_path, "r") as zip_ref:
-                    zip_ref.extractall(extract_to)
-                    extracted_files = zip_ref.namelist()
-            else:
-                raise InvalidRequestException(
-                    f"Unsupported archive format: {archive_path.suffix}"
-                )
-
-            return extracted_files
-
-        except Exception as e:
-            handle_file_error("extract", str(archive_path), e)
-
-
-class FileSearchService:
-    """Service for searching files"""
-
-    def __init__(
-        self, validation_service: FileValidationService, info_service: FileInfoService
-    ):
-        self.validation_service = validation_service
-        self.info_service = info_service
-
-    async def search_files(
-        self,
-        server_id: int,
-        search_term: str,
-        db: Session,
-        search_in_content: bool = False,
-        file_type: Optional[str] = None,
-        max_results: int = 100,
-    ) -> Dict[str, Any]:
-        """Search for files by name and optionally content
-
-        Args:
-            server_id: ID of the server to search in
-            search_term: Term to search for in filenames and/or content
-            search_in_content: Whether to search inside file contents
-            file_type: Optional file type filter
-            max_results: Maximum number of results to return
-            db: Database session (required for security validation)
-
-        Returns:
-            Dictionary containing search results and metadata
-        """
-        # Validate database session for security-critical operations
-        if db is None:
-            raise InvalidRequestException(
-                "Database session is required for file search operations"
-            )
-
-        # Validate server
-        server = await self.validation_service.validate_server_exists(server_id, db)
-        server_path = Path(server.directory_path)
-        self.validation_service.validate_server_directory(server_path)
-
-        start_time = datetime.now()
-        results = []
-
-        # Search by filename
-        filename_results = await self._search_by_filename(
-            server_path, search_term, file_type, max_results
-        )
-        results.extend(filename_results)
-
-        # Search in content if requested
-        if search_in_content and len(results) < max_results:
-            content_results = await self._search_file_content(
-                server_path, search_term, file_type, max_results - len(results)
-            )
-            results.extend(content_results)
-
-        search_time = (datetime.now() - start_time).total_seconds()
-
-        return {
-            "results": results[:max_results],
-            "total_found": len(results),
-            "search_time_seconds": round(search_time, 3),
-            "search_term": search_term,
-            "searched_content": search_in_content,
-        }
-
-    async def _search_by_filename(
-        self,
-        server_path: Path,
-        search_term: str,
-        file_type: Optional[str],
-        max_results: int,
-    ) -> List[Dict[str, Any]]:
-        """Search files by filename"""
-        results: list[dict[str, Any]] = []
-
-        for file_path in server_path.rglob("*"):
-            if len(results) >= max_results:
-                break
-
-            if search_term.lower() in file_path.name.lower():
-                try:
-                    file_info = await self.info_service.get_file_info(
-                        file_path, server_path
-                    )
-                    if not file_type or file_info["type"] == file_type:
-                        file_info["match_type"] = "filename"
-                        results.append(file_info)
-                except Exception:
-                    continue
-
-        return results
-
-    async def _search_file_content(
-        self,
-        server_path: Path,
-        search_term: str,
-        file_type: Optional[str],
-        max_results: int,
-    ) -> List[Dict[str, Any]]:
-        """Search in file content"""
-        results: list[dict[str, Any]] = []
-
-        for file_path in server_path.rglob("*"):
-            if len(results) >= max_results:
-                break
-
-            if file_path.is_file() and self.validation_service._is_readable_file(
-                file_path
-            ):
-                try:
-                    file_info = await self.info_service.get_file_info(
-                        file_path, server_path
-                    )
-                    if file_type and file_info["type"] != file_type:
-                        continue
-
-                    # Read file content and search
-                    async with aiofiles.open(
-                        file_path, mode="r", encoding="utf-8", errors="ignore"
-                    ) as f:
-                        content = await f.read()
-                        if search_term.lower() in content.lower():
-                            file_info["match_type"] = "content"
-                            results.append(file_info)
-
-                except Exception:
-                    continue
-
-        return results
 
 
 class FileManagementService:
@@ -1307,3 +605,22 @@ class FileManagementService:
 
 # Global file management service instance
 file_management_service = FileManagementService()
+
+
+# Backwards-compatibility re-exports.
+#
+# These symbols used to live in this module and are still imported and
+# patched (``@patch("app.files.application.management.<symbol>")``) by
+# existing tests and downstream callers. Keep them exposed so the split
+# is purely structural.
+__all__ = [
+    "EncodingHandler",
+    "FileBackupService",
+    "FileInfoService",
+    "FileManagementService",
+    "FileOperationService",
+    "FileSearchService",
+    "FileValidationService",
+    "file_management_service",
+    "settings",
+]
